requirements.txt
fastapi>=0.121.1
uvicorn[standard]>=0.38.0
httpx>=0.28.1
pydantic>=2.12.5
python-dotenv>=1.1.1
tenacity>=9.0.0
pytest>=9.0.1
pytest-cov>=7.0.0
pytest-asyncio>=1.3.0
pytest-httpx>=0.36.0
flake8>=7.3.0

